+++ src/components/buttons/button/Button.svelte
@@ -1,10 +1,10 @@
 <svelte:options customElement="khao-button" />
 
 <script lang="ts">
-  import { type ButtonPriority } from "./types/ButtonPriority";
-  import { type ButtonSize } from "./types/ButtonSize";
-  import Icon from "../../icons/Icon.svelte";
-  import { type IconName } from "../../icons/IconName";
+  import { type ButtonPriority, buttonPriorityDefault } from "../types/ButtonPriority";
+  import { type ButtonSize, buttonSizeDefault } from "../types/ButtonSize";
+  import Icon from "../../../icons/Icon.svelte";
+  import { type IconName } from "../../../icons/IconName";
 
   function setCustomStyles(customBGColor: string, customColor: string): string {
     var customStyles =
@@ -26,8 +26,8 @@
   export let rel: string | null = null;
   export let target: string | null = null;
 
-  export let priority: ButtonPriority = "primary";
-  export let size: ButtonSize = "medium";
+  export let priority: ButtonPriority = buttonPriorityDefault;
+  export let size: ButtonSize = buttonSizeDefault;
 
   export let customBGColor: string = "";
   export let customColor: string = "";
@@ -49,14 +49,14 @@
 >
   {label}
   {#if iconName !== ""}
-    <div class="icon icon-size-{size}">
+    <div class="icon">
       <Icon {iconName} />
     </div>
   {/if}
 </a>
 
 <style>
-  @import "reset.css";
+  @import "../reset.css";
 
   :host {
     --khao-button-container-color: var(--khao-sys-color-container);
@@ -72,11 +72,9 @@
     --khao-button-leading-space: var(--khao-sys-size-regular-5);
     --khao-button-trailing-space: var(--khao-sys-size-regular-5);
 
+    --khao-button-icon-size: var(--khao-sys-size-regular-5);
     --khao-button-icon-color: var(--khao-sys-color-on-container);
     --khao-button-icon-leading-space: var(--khao-sys-size-regular-2);
-    --khao-button-icon-size-compact: var(--khao-sys-size-regular-4);
-    --khao-button-icon-size-medium: var(--khao-sys-size-regular-5);
-    --khao-button-icon-size-large: var(--khao-sys-size-regular-6);
     
     --khao-button-height-compact: var(--khao-sys-size-regular-8);
     --khao-button-width-compact: 120px;
@@ -189,20 +187,7 @@
   .icon {
     stroke: var(--khao-button-icon-color);
     fill: var(--khao-button-icon-color);
-  }
-
-  .icon-size-compact {
-    height: var(--khao-button-icon-size-compact);
-    width: var(--khao-button-icon-size-compact);
-  }
-
-  .icon-size-medium {
-    height: var(--khao-button-icon-size-medium);
-    width: var(--khao-button-icon-size-medium);
-  }
-
-  .icon-size-large {
-    height: var(--khao-button-icon-size-large);
-    width: var(--khao-button-icon-size-large);
+    height: var(--khao-button-icon-size);
+    width: var(--khao-button-icon-size);
   }
 </style>