+++ src/components/buttons/types/ButtonPriority.ts
@@ -1,1 +1,5 @@
-export type ButtonPriority = "primary" | "secondary" | "tertiary";
+export const buttonPriorities = ["primary", "secondary", "tertiary"];
+
+export type ButtonPriority = (typeof buttonPriorities)[number];
+
+export const buttonPriorityDefault: ButtonPriority = "primary";