+++ src/components/widgets/shareButtons/ShareButtons.svelte
@@ -1,8 +1,8 @@
 <svelte:options customElement="khao-share-buttons" />
 
 <script lang="ts">
-  import Button from "../buttons/Button.svelte";
-  import openInNewWindow from "../buttons/utils/openInNewWindow";
+  import Button from "../../buttons/button/Button.svelte";
+  import openInNewWindow from "../../buttons/utils/openInNewWindow";
 
   function isAbleToPrint(): boolean {
     return window.print !== undefined;