+++ vite.split.config.ts
@@ -6,14 +6,15 @@ export default defineConfig({
   build: {
     rollupOptions: {
       input: [
-        "src/components/ads/Ad.svelte",
-        "src/components/buttons/Button.svelte",
-        "src/components/buttons/IconButton.svelte",
-        "src/components/buttons/PayPalDonateButton.svelte",
-        "src/components/buttons/SteadyButton.svelte",
-        "src/components/widgets/ShareButtons.svelte",
-        "src/components/widgets/SocialButtons.svelte",
-        "src/components/map/MarkerPin.svelte",
+        "src/components/ads/ad/Ad.svelte",
+        "src/components/buttons/button/Button.svelte",
+        "src/components/buttons/iconButton/IconButton.svelte",
+        "src/components/buttons/payPalDonateButton/PayPalDonateButton.svelte",
+        "src/components/buttons/steadyButton/SteadyButton.svelte",
+        "src/components/cards/card/Card.svelte",
+        "src/components/map/markerPin/MarkerPin.svelte",
+        "src/components/widgets/shareButtons/ShareButtons.svelte",
+        "src/components/widgets/socialButtons/SocialButtons.svelte",
       ],
       output: {
         dir: "dist/split/",